Refactored tests - moved to seperate file
